% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2014}
\alias{accident_2014}
\title{accident_2014}
\format{
3 data frames with varying number of observartions and variables
\describe{
\item{STATE}{Numeric code for US state}
\item{MONTH}{Month when accident occured}
\item{YEAR}{Year when accident occured}
\item{and many more}{...}
}
}
\usage{
accident_2014
}
\description{
Documentation of accident_2014 data
}
\examples{
 data(accident_2014)
fars_map_state(1, 2014)
fars_read_years(2014)
}
\keyword{datasets}
