% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2015}
\alias{accident_2015}
\title{accident_2015}
\format{
3 data frames with varying number of observartions and variables
\describe{
\item{STATE}{Numeric code for US state}
\item{MONTH}{Month when accident occured}
\item{YEAR}{Year when accident occured}
\item{and many more}{...}
}
}
\usage{
accident_2015
}
\description{
Documentation of accident_2015 data
}
\examples{
data(accident_2015)
fars_map_state(1, 2015)
fars_read_years(2015)
}
\keyword{datasets}
