% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctest-process.R
\name{doctest}
\alias{doctest}
\title{Create doctests from roxygen tags}
\usage{
doctest()
}
\value{
The doctest roclet
}
\description{
To use this, add
}
\details{
```
Roxygen: list(roclets = c("collate", "rd", "namespace", "doctest::doctest"))
```

to your package DESCRIPTION file. Then run [roxygen2::roxygenize()] from the
command line. Or, run `roxygen2::roxygenize(roclets = "doctest::doctest")`.
}
