% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{doctestExample-tag}
\alias{doctestExample-tag}
\alias{@doctestExample}
\title{Add an example from a file}
\description{
\verb{@doctestExample path/to/file.R} is a drop-in replacement for
\verb{@example path/to/file.R}. It doesn't add the contents of \code{file.R} to
the test.
}
\details{
If you have complex examples you may want to store them separately.
Roxygen2 uses the \verb{@example} tag for this. \verb{@doctestExample} does the
same: it adds the contents of its file to the resulting example.
Suppose \code{man/R/example-code.R} contains the line:

\if{html}{\out{<div class="sourceCode">}}\preformatted{2 + 2
}\if{html}{\out{</div>}}

Then the following roxygen:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @doctest
#'
#' @expect equal(2)
#' 1 + 1
#' @doctestExample man/R/example-code.R
}\if{html}{\out{</div>}}

will generate an example like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{1 + 1
2 + 2
}\if{html}{\out{</div>}}

At present, \verb{@doctestExample} doesn't add any code to the tests.

\verb{@doctestExample} was added in doctest 0.3.0.
}
