% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-docs.R
\name{expect}
\alias{expect}
\alias{@expect}
\title{Create an expectation}
\description{
\verb{@expect} creates an expectation for your example code.
}
\details{
Use \verb{@expect} to create a testthat expectation.

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#'
#'x <- 2 + 2
#' @expect equals(x, 4)
#'
#' f <- function () warning("Watch out")
#' @expect warning(f())
}\if{html}{\out{</div>}}

Don't include the \code{expect_} prefix.

Use a dot \code{.} to refer to the following expression:

\if{html}{\out{<div class="sourceCode">}}\preformatted{@expect equals(., 4)
2 + 2
@expect length(., 10)
rnorm(10)
}\if{html}{\out{</div>}}
}
