% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer identifier of the state in accidents data.}

\item{year}{year for which accident data is desired. Must/should be numeric.}
}
\value{
NULL, graphic of the state will be plotted.
}
\description{
\code{fars_map_state} plots where the accidents of a chosen year and state have
happend.
}
\note{
If !(year \%in\% 2013:2015) the function will throw an error
}
\examples{
fars_map_state(1, 2013)
}
