% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{csv to tibble conversion}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name or path of dataset saved in csv. Must be of class character.}
}
\value{
a tibble data frame
}
\description{
\code{fars_read} takes a name of package-internal rda dataset as input and
converts it to a tibble data frame.
}
\note{
If your input refers to a file that does not exist in the package, the function
will fail.
}
\examples{
doctest:::fars_read('accident_2013')
}
