% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Retrieve MONTHS, years accident data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of years for which accident data should be retrieved.
Must/should be a numeric vector.}
}
\value{
a list of length length(years) of nX2 tibble data frames
}
\description{
\code{fars_read_years} takes years as an argument and outputs the months and years of
the corresponding accident data.
}
\note{
If any(!(years \%in\% 2013:2015)) the function will issue a warning and output a
NULL value.
}
\examples{
fars_read_years(2013:2015)
}
