% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count accidents}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of years for which accident data should be retrieved.
Must/should be a numeric vector.}
}
\value{
tibble data frame of accident counts
}
\description{
\code{fars_summarize_years} counts the accidents by year and month of the accident datasets.
}
\note{
If any(!(years \%in\% 2013:2015)) the function will issue a warning and ignore the
incorrect input. Will throw an error if all inputs are incorrect.
}
\examples{
fars_read_years(2013:2015)
}
