% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Path to accident data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year for which accident data is desired. Must/should be numeric.}
}
\value{
a string of the name of the accident data
}
\description{
\code{make_filename} takes year as an argument and creates the filename for
the accident data of the corresponding year
}
\note{
If ! (year \%in\% 2013:2015) the function will create a string (or NAs)
for which no accident dataset exists.
}
\examples{
doctest:::make_filename(2013)
}
