% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-docs.R
\name{pause-tag}
\alias{pause-tag}
\alias{@pause}
\alias{@resume}
\alias{resume-tag}
\title{Exclude example code from a test}
\description{
\verb{@pause} excludes code from a test until the next tag.
Use \verb{@resume} to restart including code without creating an expectation.
}
\details{
Use \verb{@pause} to avoid redundant or noisy code:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @doctest
#'
#' @expect equal(0)
#' sin(0)
#'
#' @pause
#' curve(sin(x), 0, 2 * pi)
#'
#' @expect equal(1)
#' cos(0)
}\if{html}{\out{</div>}}

Remember that the main purpose of examples is to document your package for
your users. If your code is getting too different from your example, consider
splitting it off into a proper test file. You can do this by renaming it and
deleting the \verb{Generated by doctest} comment.
}
