% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-docs.R
\name{skipTest-tag}
\alias{skipTest-tag}
\alias{@skipTest}
\alias{@resumeTest}
\title{Exclude example code from a test}
\description{
\verb{@skipTest} excludes code from a test until the next tag.
Use \verb{@resumeTest} to restart including code without creating an expectation.
}
\details{
Use \verb{@skipTest} to avoid redundant or noisy code:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#'
#' @expect equal(0)
#' sin(0)
#'
#' @skipTest
#' curve(sin(x), 0, 2 * pi)
#'
#' @expect equal(1)
#' cos(0)
}\if{html}{\out{</div>}}

Remember that the main purpose of examples is to document your package for
your users. If your code is getting too different from your example, consider
splitting it off into a proper test file. You can do this by renaming it and
deleting the \verb{Generated by doctest} comment.
}
