% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-docs.R
\name{skiptest}
\alias{skiptest}
\alias{@skiptest}
\alias{@unskip}
\title{Exclude example code from a test}
\description{
\verb{@skiptest} excludes the following code from a test. \verb{@unskip} stops excluding
code.
}
\details{
Use these tags to avoid redundant or noisy code:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#'
#' @expect equal(., 0)
#' sin(0)
#'
#' @skiptest
#' curve(sin(x), 0, 2 * pi)
#' @unskip
#'
#' @expect equal(., 1)
#' cos(0)
}\if{html}{\out{</div>}}

Remember that the main purpose of examples is to document your package for
your users. If your code is getting too different from your example, consider
splitting it off into a proper test file.
}
