% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{snap-tag}
\alias{snap-tag}
\alias{@snap}
\title{Create a snapshot test}
\description{
\verb{@snap} creates a
\href{https://testthat.r-lib.org/articles/snapshotting.html}{snapshot test}
for your example. It is shorthand for \verb{@expect snapshot()}.
}
\details{
Often, examples show complex output to the user. If this output changes,
you want to check that it still "looks right". Snapshot tests help by
failing when the output changes, and allowing you to review and approve
the new output.

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @doctest
#'
#' @snap
#' summary(lm(Petal.Width ~ Species, data = iris))
}\if{html}{\out{</div>}}
}
\seealso{
Other expectations:
\code{\link{expect-tag}},
\code{\link{expectRaw-tag}}
}
\concept{expectations}
