% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-docs.R
\name{test}
\alias{test}
\alias{@test}
\title{Start a new test}
\description{
\verb{@test} starts a new test.
}
\details{
By default, a test labelled "Example: <object name>" is created. Use
\verb{@test} to create separate tests within a single example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#'
#' x <- 1
#' @expect equal(., x)
#' abs(x)
#'
#' @test Negative numbers
#' x <- -1
#' @expect equal(., -x)
#' abs(x)
}\if{html}{\out{</div>}}
}
