% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-docs.R
\name{testComments}
\alias{testComments}
\alias{@testComments}
\title{Write expectations in comments}
\description{
\verb{@testComments} lets you write expectations in comments.
This may be useful if you have complex examples with \code{if} statements or
\code{for} loops
}
\details{
Doctests like this won't work, because the test tags split up the example
code so that roxygen can't parse it.

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#'
#' if (x > 0) \{
#'   @expect gt(x, 0)
#'   print("x is positive")
#' \} else \{
#'   @expect lt(x, 0)
#'   print("x is negative")
#' \}
}\if{html}{\out{</div>}}

As an alternative, put \verb{@testComments} in your examples section and write
expectations in comments:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#' @testComments
#' if (x > 0) \{
#'   # expect gt(x, 0)
#'   print("x is positive")
#' \} else \{
#'   # expect lt(x, 0)
#'   print("x is negative")
#' \}
}\if{html}{\out{</div>}}

Expectations in comments match the format of \link{expect} tags, but with a
comment character \verb{#} in place of the \code{@}.

Comments should be on their own line, with no other code:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' # Wrong:
#' x <- 2 + 2 # expect equal(x, 4)
#'
#' # Right:
#' x <- 2 + 2
#' # expect equal(x, 4)
#'
#' # Right:
#' # expect equal(., 4)
#' 2 + 2
}\if{html}{\out{</div>}}
}
