% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{package}
\name{doctest-package}
\alias{doctest}
\alias{doctest-package}
\title{Write testthat tests and examples together}
\description{
Doctest lets you combine tests and examples. For more information, see
\link{@doctest} and \link{@expect}.
}
\details{
\subsection{Example}{

Here's some \href{https://roxygen2.r-lib.org}{roxygen} documentation for a function:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
#' Safe mean
#' 
#' @param x Numeric vector
#' @return The mean of `x`
#' 
#' @doctest
#' @expect equal(2)
#' safe_mean(1:3)
#' 
#' @expect warning("not numeric")
#' safe_mean("a")
#'
#' @expect warning("NA elements")
#' safe_mean(c(1, NA))
safe_mean <- function (x) \{
  if (! is.numeric(x)) warning("x is not numeric")
  if (any(is.na(x))) warning("x contains NA elements")
  mean(x)
\}
}\if{html}{\out{</div>}}

Instead of an \verb{@examples} section, we have a \verb{@doctest} section.

This will create tests like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{# Generated by doctest: do not edit by hand
# Please edit file in R/<text>

test_that("Doctest: safe_mean", \{
  # Created from @doctest for `safe_mean`
  # Source file: <text>
  # Source line: 7
  expect_equal(safe_mean(1:3), 2)
  expect_warning(safe_mean("a"), "not numeric")
  expect_warning(safe_mean(c(1, NA)), "NA elements")
\})
}\if{html}{\out{</div>}}

The .Rd file will be created as normal, with an example section like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\examples\{
safe_mean(1:3)
safe_mean("a")
safe_mean(c(1, NA))
\}
}\if{html}{\out{</div>}}
}

\subsection{Usage}{

You can install the development version of doctest like this:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{devtools::install("hughjonesd/doctest")
}\if{html}{\out{</div>}}

To use doctest in your package, alter its DESCRIPTION file to add the
\code{dt_roclet} roclet to roxygen:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{Roxygen: list(roclets = c("collate", "rd", "namespace", 
              "doctest::dt_roclet")) 
}\if{html}{\out{</div>}}

Add doctest to "Suggests" as a dependency:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{usethis::use_dev_package("doctest", type = "Suggests", 
                         remote = "hughjonesd/doctest")
}\if{html}{\out{</div>}}

Then run \code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}} or \code{\link[devtools:document]{devtools::document()}} from the
command line.

Doctest is \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hughjonesd.github.io/doctest/}
  \item Report bugs at \url{https://github.com/hughjonesd/doctest/issues}
}

}
\author{
\strong{Maintainer}: David Hugh-Jones \email{davidhughjones@gmail.com}

}
