% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctest-package.R
\docType{package}
\name{doctest-package}
\alias{doctest}
\alias{doctest-package}
\title{Write testthat files using roxygen tags}
\description{
To use this, add

\if{html}{\out{<div class="sourceCode r">}}\preformatted{Roxygen: list(roclets = c("collate", "rd", "namespace", "doctest::doctest"))
}\if{html}{\out{</div>}}

to your package DESCRIPTION file. Then run \code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}} or
\code{\link[devtools:document]{devtools::document()}} from the command line.
}
\author{
\strong{Maintainer}: David Hugh-Jones \email{davidhughjones@gmail.com}

}
\keyword{internal}
