% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctest-package.R
\docType{package}
\name{doctest-package}
\alias{doctest}
\alias{doctest-package}
\title{doctest: Summaries for the accident data}
\description{
package includes the `fars_map_state` and `fars_summarize_years` functions to creat plots and summaries of the accident data.
}
\author{
\strong{Maintainer}: MA BO \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
