% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-docs.R
\name{doctest-tag}
\alias{doctest-tag}
\alias{@doctest}
\title{Start a  doctest}
\description{
\verb{@doctest} starts a doctest.
}
\details{
Use \verb{@doctest} where you would usually use \verb{@examples}. Then add
\link{@expect} and \link{@expectRaw} tags beneath it to create expectations.

You can have more than one \verb{@doctest} tag. Each doctest will create
a new test, but they will all be merged into a single Rd example.

By default, a test labelled "Example: <object name>" is created. You
can put a different label after \verb{@doctest}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @doctest Positive numbers
#'
#' x <- 1
#' @expect equal(x)
#' abs(x)
#'
#' @doctest Negative numbers
#' x <- -1
#' @expect equal(-x)
#' abs(x)
}\if{html}{\out{</div>}}
}
