% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{doctest}
\alias{doctest}
\title{Create the doctest roclet}
\usage{
doctest()
}
\value{
The doctest roclet
}
\description{
You can use this in your package DESCRIPTION like this:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{Roxygen: list(roclets = c("collate", "rd", "namespace", "doctest::doctest"))
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
roxygen2::roxygenize(roclets = "doctest::doctest")
}
}
