% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctest-process.R
\name{doctest}
\alias{doctest}
\title{Create doctests from roxygen tags}
\usage{
doctest()
}
\value{
The doctest roclet
}
\description{
Create doctests from roxygen tags
}
\details{
To use this, add

\if{html}{\out{<div class="sourceCode r">}}\preformatted{Roxygen: list(roclets = c("collate", "rd", "namespace", "doctest::doctest"))
}\if{html}{\out{</div>}}

to your package DESCRIPTION file. Then run \code{\link[roxygen2:roxygenize]{roxygen2::roxygenize()}} from the
command line. Or, run \code{roxygen2::roxygenize(roclets = "doctest::doctest")}.
}
