% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{doctest}
\alias{doctest}
\title{Create doctests from roxygen tags}
\usage{
doctest()
}
\value{
The doctest roclet
}
\description{
Create doctests from roxygen tags
}
\examples{
\dontrun{
roxygen2::roxygenize(roclets = "doctest::doctest")
}
}
