% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-docs.R
\name{expect-tag}
\alias{expect-tag}
\alias{@expect}
\title{Create an expectation}
\description{
\verb{@expect} creates an expectation for your example code.
}
\details{
Use \verb{@expect} to create a testthat expectation.

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#'
#' @expect equals(4)
#' 2 + 2
#'
#' f <- function () warning("Watch out")
#' @expect warning()
#' f()
}\if{html}{\out{</div>}}

The next expression will be inserted as the first
argument to the \verb{expect_*} call.

Don't include the \code{expect_} prefix.

If you want to include the expression in a different
place or places, use a dot \code{.}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{@expect equals(., rev(.))
c("T", "E", "N", "E", "T")
}\if{html}{\out{</div>}}
}
