% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-docs.R
\name{expectRaw-tag}
\alias{expectRaw-tag}
\alias{@expectRaw}
\title{Create an expectation as-is}
\description{
\verb{@expect} creates an expectation for your example code.
}
\details{
\verb{@expectRaw} creates a testthat expectation. Unlike \link{@expect}, it doesn't
insert the next expression:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#' @examples
#'
#'x <- 2 + 2
#' @expectRaw equals(x, 4)
#'
#' f <- function () warning("Watch out")
#' @expectRaw warning(f())
}\if{html}{\out{</div>}}

Don't include the \code{expect_} prefix.
}
