% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_doctests}
\alias{test_doctests}
\title{Test doctests in a package}
\usage{
test_doctests(path = ".", ...)
}
\arguments{
\item{path}{Path to package}

\item{...}{Passed to \code{\link[testthat:test_package]{testthat::test_local()}}.}
}
\value{
The result of \code{\link[testthat:test_package]{testthat::test_local()}}.
}
\description{
This is a utility function to run doctests in a local source package.
It calls \code{\link[testthat:test_package]{testthat::test_local()}}.
}
\examples{
\dontrun{
  test_doctests()
}
}
